- removed additional context run for kafka tests for schema enabled, not tests are run on configuration with enabled schema - merged keycloack tests contexts, not we have two contexts run with and w\o realm - removed 2 contexts from Oracle tests (we tested Oracle xe11 eventhough its default image already) also removed additional context with initScript, now all tests are run with initial script setting - added missing modules to BOM
